Fix coverage with esm
